完善视频流提交

+++ modules/wvp/src/main/java/com/cdzeroly/wvp/domain/bo/ChannelToGroupByGbDeviceParam.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 @Data
 public class ChannelToGroupByGbDeviceParam {
-    private List<Integer> deviceIds;
+    private List<Long> deviceIds;
     private String parentId;
     private String businessGroup;
 }
